Changed set_sound_program function to set status directly
The current way of updating status relies on a worker thread that listens for UDP status messages from the musiccast receiver. The UDP subscription is renewed periodically by the get_status function.
This status update methodology doesn't seem to work for a sound program change because of the way the device responds to a sound program change - while a UDP status message is sent after a sound program change, it doesn't seem to contain any meaningful info, just something like
 {"main":{"status_updated":true},"device_id":"xxxxxxxxxxxx"}. Because of this, the sound program state is not updated immediately and is not updated until the next time get_status is called as part of its resubscribe loop.
I put in a quick and dirty workaround which just sets the sound_program state directly. However this only will only reflect status instantly if the state is updated via HA - for updates from elsewhere we still have to wait for the next resubscribe loop.